third_party/blink/web_tests/images/imagemap-scroll-expected.txt

This tests to be sure that focusing an area element triggers scrolling and removing focus from it does not.

PASS: Document is starting scrolled to top.

PASS: Focusing area element caused the image to scroll into view.

PASS: Document is scrolled to top once again.

PASS: Document is still scrolled to top after removing focus from area element.

TEST COMPLETE
